Added startup confirmation from subsystems using I2C
When AOCS or payloads transmit a 0x1 (a temporary placeholder that may or may not be used in the final version) to the controller via I2C it is interpreted as that they've finished their startup.
When the controller has received startup confirmation it enters the next state, P2DetectionState.